Revert "Skip costing group by plans for empty tables"
Summary:
It turns out that division by zero on empty tables is a problem specific
to MyRocks. It turns out that storage engines that don't support
HA_STATS_RECORDS_IS_EXACT should never return 0 for number of records.
Otherwise, the SQL layer may assume that the table is empty and elide
reading from it, which would be incorrect if the table actually
contained data.

This is indicated in this comment:

  /*
    The number of records in the table.
      0    - means the table has exactly 0 rows
    other  - if (table_flags() & HA_STATS_RECORDS_IS_EXACT)
               the value is the exact number of records in the table
             else
               it is an estimate
  */
  ha_rows records;

To fix, copy InnoDB's behaviour and always return >= 1.

Squash with: D35825006

Reviewed By: luqun

Differential Revision: D37399246

fbshipit-source-id: 0185b05a2c0e2cfcd3fdbb89f0aec6052ecfede5