Remove multifactor flag from the default client flags
Summary:
The MFA client flag in 8.0.28 conflicts with the lz4_stream compression
flag in 8.0.23. While 8.0.28 removed the lz4_stream flag, it will fail
when connecting to 8.0.23 unless the compression_lib is set. But not
all applications set compression_lib.

Removing this flag from the default client flag will allow these
application to connect to 8.0.23. Once everything is on 8.0.28, this
patch can be dropped.

Squash with null

Reviewed By: lth

Differential Revision: D37718979