Fix promotion of a specific leader
Summary:
This is part of a series of diffs to fix raft specific MTR tests.

The problem was that our script to promote a specific leader wasn't actually promoting a specific leader. It was only starting elections. The elected leader could be anyone.

In many of our tests, we rely on specific leaders to be elected. As a result, many tests fail. The test plan highlights one of the test cases that it fixes.

It looks like the cause of this issue might have been a bad merge in D33193348.

Reviewed By: yichenshen

Differential Revision: D38181610

fbshipit-source-id: 513caa4475d87916c038931a6fcf5888bece2018