fix the issue of 0 io_write issue when operating on partition table
Summary: For partition tables, there are missing some info in the performance schema table. They are io_write_bytes, io_write_requests, io_read_bytes and io_read_requests.The current return value of these field are 0, which is not the correct value. In this diff, the issue of 0 return value is fixed.

Reviewed By: luqun

Differential Revision: D38227596

fbshipit-source-id: da1d01c17e2689ea8ef51276bd67a34170a6f8ba