Add size limit for intrinsic tmp table in rocksdb
Summary:
This diff adds a size limit to intrinsic tmp table in rocksdb.  We will validate the size of intrinsic tmp table for each write_row/update_row and fail the query if it exceeds the threshold.

This diff follows the changes in the innodb diff: D25237256. It uses the infrastructure in innodb diff to maintain and update the session and global tmp disk usage. We increase the tmp disk usage when writing row and updating row, decrease the tmp disk usage when we delete intrinsic tables. The increased value are obvious. To calculate the value we need minus when deleting intrinsic tables, we add two more variables here:

- open_tmp_tables : We use it to track the active intrinsic tmp tables. We increase it when creating a new intrinsic tmp table and decrease it when deleting one. Let's assume a query will create intrinsic tmp table, when open_tmp_tables is empty, it means a query has finished. It's the time that we need to decrease the disk usage.

- total_tmp_table_size : It's the amount of value we need to update when we want to decrease the disk usage.

We use **max_tmp_disk_usage** as the threshold of global intrinsic tmp table usage.

The batch processing of global disk usage has been handled by innodb diff, so we don't do additional work here.

Reviewed By: Pushapgl

Differential Revision: D38514012

fbshipit-source-id: b79c35b7e84b04f3596f08009d2f785a350229c5