MyRocks TTL timestamps in binlogs
Summary:
Maintaining two TTL timestamps in the binlog:
1. Compaction timestamp which is used to calculate row expiry during
compaction. This is written once every
rocksdb_binlog_ttl_compaction_ts_interval_secs secs to the binlog in the
metadata event of a transaction.
2. Read filtering timestamp which is used to get a consistent view and
calculate row expiry on a per transaction basis. This is written with
every transaction in the binlog in the metadata event.

Since these timestamps are written in the binlog the secondaries can
read it and use it for its own row expiry calculation. This provides
consistent reads across primary and secondaries as timestamps used for
TTL related calculations are serialized along with transactions in the
binlog.

These timestamps are derived from HLC so they're event driven (clock is
bumped once per binlogged transaction) and automatically take care of
clock drifts and synchronization between primary and secondaries.

Reviewed By: hermanlee

Differential Revision: D33912509