Make STOP SLAVE work during LOCK INSTANCE FOR BACKUP
Summary: When the LOCK INSTANCE FOR BACKUP runs, the "stop slave" command fails to run until the unlock of instance. This will prevent the slave replica to be promoted to primary. This diff is to bypass the check of lock by turn on the global variable to allow the "stop slave" command to run during this scenario.

Reviewed By: luqun, Pushapgl

Differential Revision: D38845865

fbshipit-source-id: 43f9940fda8a60442b7f698729b3563f8a045ab9