Use engine binlog pos to initialize gtid_executed on crash recovery
Summary:
The server maintains max of all recovered gtids and the min of
all recovered binlog position from all engines. On a host restart, there
can be a case where rocksdb's transaction log gets truncated more than
innodb's. In this case if we use the gtid table (which is an innodb
table) to init gtid_executed we'll end up overestimating the
gtid_executed which can lead to transactions begin skipped. Ideally we
should not rely on gtid table at all and just use min binlog position to
construct the gtid_executed. This is what was happening in 5.6 as well
but we missed porting it correctly to 8.0.

In this patch we're skipping gtid table and using the min recovered
binlog pos to construct gtid_executed on server recovery.

Reviewed By: yichenshen

Differential Revision:
D38887898

Signed-off-by: Yichen <yichenshen@fb.com>

fbshipit-source-id: e298304459b013044c6f22d052724ced0641abff