Fix the MTR test after cherry-picking change from 8.0.23 to 8.0.28
Summary:
D38845865 add a global variable 'skip_backup_lock_for_stop_replica' in 8.0.23 as a work around of promotion failure. (Task for context: T129393799) We have the same issue in 8.0.28. So the same change was cherry-picked to 8.0.28. Looks like we have two more tests to fix for 8.0.28. These two tests passed in 8.0.23.

Squash with D38845865

Reviewed By: hermanlee

Differential Revision: D39200404

fbshipit-source-id: bf3fb0c9b8aa92396c63af79fb48c5f76f69a820