Fix range scans not to overestimate io cost, and not to eliminate the index from fallback plan
Summary:
MySQL 8.0 optimizer has a behavior called "range_uses_more_keyparts" to disable the "ref" plan for the index in favor of the "range" plan, if the same or more key parts can be used for the same index. 5.6 has similar behavior, but 8.0 eliminates fallback possibilities -- if range for the index is excluded from the plan later (e.g. estimated cost gets bigger than other plans), the index is not considered as a ref plan either. This may cause a plan regression
to pick a suboptimal index. Relevant bug report is https://bugs.mysql.com/bug.php?id=104083

For example, there are indexes i1 (c1) and i2 (c1, c2), and issuing a query with “WHERE c1=1 AND c2 BETWEEN 1 and 2”. Range plan with i2 is likely the best plan. However, 8.0 may pick ref with i1. This can happen when range scan cost with i2 is estimated higher than ref with i1. Ref with i2 should be more efficient than i1, but "range_uses_more_keyparts" excludes the ref i2 plan.

The overestimating range scan issue exists in MySQL 5.6 as well, but MySQL 8.0 excludes the chance to pick the index for any other plan (notably ref plan), so plan regression impact is much worse. Obviously this is not a desired behavior.
One of the notable reasons why range plan cost is estimated higher than ref is that ref caps io cost estimates but range does not.
MySQL optimizer is cost based, and the cost mainly consists of io cost and cpu cost.
Typically, the range query optimizer plan calls handler::read_cost() to estimate io cost. If there are many rows that are estimated to read, the read_cost() can be a lot higher than the ref plan. This is because ref plan caps io cost by calling find_worst_seeks(), while range plan does not cap anything. This makes the optimizer pick ref rather than range if the scan range is estimated wider (larger than 10%). In MySQL 5.6, there is still a chance to pick a range plan for the index, if the optimizer picks a ref plan for the index. Optimizer later changes ref to range because it can use more key parts (“uses_more_keyparts"). In MySQL 8.0, the fallback logic is gone and the index is excluded from other than range scan, it ends up picking a different index, which is worse for execution.

This diff adds a session variable optimizer-fix-range-cost-row-threshold. If the number of estimated total rows of the table is bigger than the value, optimizer calls find_worst_seeks() to cap io cost, and the index is still considered as a candidate even if the range cost is estimated higher. Eventually range plan may be picked by "uses_more_keyparts".

Reviewed By: luqun

Differential Revision: D39678105

fbshipit-source-id: 9bbcac38d4d99e5350cdf5f564f5467de800f2f7