Add flag to check for materialized groups before issuing SD
Summary:
Today, when partial indexes need maintenance, an SD is always issued
blindly in order to save a read. This could potentially cause tombstones to build up
though, and the cost of performing a point lookup turns out to be not
that bad.

Add a flag to control this behaviour so that we can adjust if this becomes
a problem later.

Reviewed By: luqun

Differential Revision: D40196622

fbshipit-source-id: 5e4c114a001d6c6b75e942f4a521003e5a7c4603