Adding a flag to delay auto binlog rotation when a raft purge is active
Summary:
Purging logs in raft is very expensive because we read logs to
find the last opid for every to-be-purged binlog. Ideally we should be
storing this info in the index file like the prev gtid set but that is a
big work item. Purge can cause stalls in the commit pipeline when auto
rotation is triggered (current binlog goes beyond max_binlog_size). To
quick fix added an atomic to indicate if a raft purge is active, if yes,
we delay to auto rotate operation in group commit, avoiding stalls at
the cost of slighly exceeding max_binlog_size.

Reviewed By: anirbanr-fb

Differential Revision: D40743288

fbshipit-source-id: f51a6541f6d92d624d7c8f80d2caf586d1de6884