Fix inefficient range tree construction from IN list
Summary:
Fixes https://bugs.mysql.com/108963

When constructing the range tree, we end up re-merging the current tree for every element of the IN list, and this causes optimization to be very slow. Fix this by switching the order of arguments, so that we always merge the new item into the existing tree instead.

I've also considered fixing `key_or` to automatically swap the order of arguments, but there is already existing swapping logic that considers which tree needs copied vs can be edited in place, and it seems unclear how to integrate with that effectively (since the two goals could be conflicting).

Reviewed By: hermanlee

Differential Revision: D40948362

fbshipit-source-id: 04ddd79b79ad0d6e9b585f56c379b5fe0aee41b4