Fix truncate_failures MTR test
Summary: Fix the MTR test for truncate_failures. The failure is caused when  restarting the mysqld after crashing the server during server truncation. When initializing rocksdb plugin, the table will be removed, which is the expected behavior. However, handlerton slot number for rocksdb is undefined at this code point. D38514012 called `get_ha_data(ha_thd())`, which will use the undefined slot number and then lead to a crash. Move the function call inside of the intrinstic tmp table condition to avoid calling it in the initialization part. The slot number will be assigned a value later so it will not cause issue after initilization.

Reviewed By: hermanlee

Differential Revision: D41292242

fbshipit-source-id: f8a7a317f99ffd4ae3920f9d303553b5491a50a6