MyRocks Clone (5/n): fix bug 109922 (SEs do not coordinate clone rollbacks on instance startup)
If InnoDB clone decides to rollback during its startup, other storage engines
are not aware of this and proceed as usual, resulting in inconsistent instance
between storage engines. Ideally this should be fixed through a 2PC-like design
where all the SEs vote to rollback the clone or to continue during the startup,
but for MyRocks clone a simpler solution suffices: make InnoDB write a special
marker file if it decides to rollback on clone startup.