Disallowing binlog commits on raft followers
Summary:
We don't want to allow any direct writes on the raft followers.
To achieve this invariant we only allow direct binlog writes (i.e.
without before_flush) when we're writing to an apply log and the thread
that's doing the write is a replication applier thread. In addition, we
don't allow before_flush to be called when we're in apply log mode.
before_flush will also fail within kudu when the current raft role is
not leader.

Checking MYSQL_BIN_LOG::is_apply_log is safe inside
MYSQL_BIN_LOG::write_transaction() because it only changes while
LOCK_log is held while switching logs.

Reviewed By: yichenshen

Differential Revision: D41032937

fbshipit-source-id: 62bd91dc1cc0e267bebcbe8be9e30524bc5070be