Fix main.myisam_explain_json_non_select_all
Summary:
This is mainly a test fix for `main.myisam_explain_json_non_select_all`. There are a few other places where we erroneously show MRR used for descending scans, and this does not address those.

The reason why EXPLAIN shows MRR used is because `test_quick_select` needs to be called again with the correct `interesting_order` flag. We don't always do this, and with the previous fix D37321652 regressed this behaviour.

Note that this only affects EXPLAIN since during execution, we will always disable use of MRR in `ReverseIndexRangeScanIterator`.

Squash with: D37321652

Reviewed By: hermanlee

Differential Revision: D41637647

fbshipit-source-id: 2b1fe71d368da0f6de6b1d87a2eed5166d4f57eb