Bug#34891365: MySQL 8.0.29+ is slower than MySQL 8.0.28 with queries using JOINS
Summary:
FB: Backport fix for performance regression

The one-row cache in EQRefIterator was disabled for queries using
streaming aggregation in the fix for bug#33674441. This was done in
order to fix a number of wrong result bugs, but it turned out to have
too big impact on the performance of many queries.

This patch re-enables the cache for the above queries, and fixes the
original wrong result bugs in a different way. There were two
underlying problems that caused the wrong results:

1) AggregateIterator::Init() does not restore the table buffers to the
state they had after the last read from the child iterator in the
previous invocation. The table buffers would have the content from the
first row in the last group that was returned from the iterator in the
previous invocation, rather than the contents of the last row read by
the child iterator, and this made the cache in EQRefIterator return
wrong values. Fixed by restoring the table buffers in
AggregateIterator::Init() if the previous invocation had modified
them.

2) When the inner tables of an outer join are null-complemented, the
table buffers are modified to contain NULL values, thereby disturbing
the cached value for any EQRefIterator reading from one of the
null-complemented tables. Fixed by making StoreFromTableBuffers()
store the actual values contained in the table buffer instead of the
null values, if the table is accessed by EQRefIterator.
LoadIntoTableBuffers() is taught to restore those values, but
additionally set the null flags on the columns after restoring them.

The hypergraph optimizer had another workaround for these wrong
results (it added a streaming step right before the aggregation). This
workaround is also removed now.

Change-Id: I554a90213cae60749dd6407e48a745bc71578e0c

Post-push fix.

Implicitly grouped queries on tables in the CSV engine failed under
ASan if the tables contained BLOBs.

The original bug fix made AggregateIterator store the contents of all
the record buffers when its child iterator was exhausted, so that they
could be restored if the AggregateIterator was reinitialized. This
interacted badly with the special handling of BLOB memory in the CSV
engine. The CSV engine has a separate MEM_ROOT for BLOB memory, and it
clears it for reuse before reading a new row. But it clears it before
it checks if it has reached EOF, so when EOF is reached, the
Field_blob objects will point to freed memory. Trying to copy that
memory into the AggregateIterator's buffer makes ASan fail.

Fixed by clearing the CSV engine's BLOB MEM_ROOT for reuse only when
there actually is a new row that can reuse the memory. This makes the
record buffer contain valid data after EOF is seen; either it has the
data from the last row, or it has empty data if the table was empty.
Then it's safe to store and restore the record buffer even after EOF.

Change-Id: I9b5051b0a16933c687677db6c4fe761d1eef7b96

Post-push fix.

Rerecord a test that failed because a query plan was changed by the
bug fix.

Change-Id: I61a1274e314f6bab7a3a83304b810fc9591c79c1

Squash with: null

Differential Revision: D53247889

fbshipit-source-id: ed9c6db8d5fd6f4f7824fd3ae6b4ce5f01eea946