MyRocks Clone (2/n): fix bug 109919 (Clone plugin does not support arbitrary length O_DIRECT files)
The existing clone plugin I/O routines handle O_DIRECT files with the assumption
that their sizes are always a multiple of a block size. This is too restrictive
because in general O_DIRECT files can be of any size and MyRocks files actually are.

Add a flag to Ha_clone_file to support arbitrary file lengths under O_DIRECT:
o_direct_uneven_file_size. Pass it to clone_client.h/cc
Client::get_aligned_buffer to allocate the buffer that bumps the length to the
next CLONE_OS_ALIGN multiple if it's not already one. Introduce a
pad_for_o_direct helper that performs this calculation for the passed length as
needed.

For writes, if O_DIRECT is enabled and the buffer length is not a multiple of
CLONE_OS_ALIGN, pad it with zeros to the next multiple. Then check the actual
bytes written, and if any padding bytes were actually written, call my_chsize to
truncate the file to the needed size.