Fix main.execution_constants
Summary: This test is failing in release because we are exhausting the parser memory limit, before hitting stack overflow. The reason is because when fdo is turned on (note, not lto), enough functions are inlined such that we end up consuming much less stack space. To resolve, use different expressions where functions are less inlined.

Reviewed By: luqun

Differential Revision: D43136864

fbshipit-source-id: df1235878f99aaa6aa7b064df7668a1319576c30