Add support for DSCP (Differentiated Services Field) to MySQL Clone Plugin for remote clone process.
MySQL Clone Plugin allows for faster data copy using multi-threaded process. Setting DSCP fields on sockets that are created using randomly selected ports is quite difficult and will always be delayed due to fetching of the socket information from MySQL processlist or other methods. With clone_dscp_on_socket MySQL Clone plugin variable we are able to set the DSCP value on the socket as soon as they are created for remote transfers.