Do not cache large allocations used for encoding table records
Summary:
The memory on `m_storage_record` is cached on every table handler. This memory is used for encoding the value slice when writing to rocksdb. This means that if there has been a large allocation due to large blobs on the table record, this memory will stick around until the handler is freed. This could be a long time, since our table handlers are cached.

The fix is to release the memory when the handler is returned to table handler cache, if it exceeds some threshold. This way we can avoid holding onto large allocations, while avoiding regressing the small allocations case.

Reviewed By: luqun

Differential Revision: D43753421