Make blackhole engine transactional
Summary:
The main motivation for this diff is that we want to be able to set `super_read_only` and `read_only`=1 very quickly. Normally, when there are ongoing transactions happening, setting the read only vars will time out. This is problematic for promotions in both semi-sync and raft.

The timeout is caused by long running transactions holding on to the read lock. In 2016, we introduced a sysvar `legacy_global_read_lock_mode` in MySQL 5.6 (D5130471) to allow us to skip acquiring the read lock when turning on read only variables. It would skip the read lock if `legacy_global_read_lock_mode=0`  This would mean that turning read only would be instant and any ongoing transactions would eventually die at commit.

However, it was deemed not safe to skip the read lock because DDL statements could still be completed despite read only vars being turned on. Instead, the team decided to implement the strategy of killing of conflicting connections (D9563242) after read only is turned on to mitigate the long running transactions problem. It is only a mitigation because after killing the transactions, we still have to wait for the rollback before the read lock is released and this can be slow and still result in a timeout.

In MySQL 8.0, transactional DDL was supported. This is great because DDL statements would be treated the same way as DML statements. So if read-only was enabled and there were an ongoing DDL statement, it will fail at commit. This means that we can safely utilize `legacy_global_read_lock_mode=0` and skip the read lock when setting read only vars.

However, the catch is that in MySQL 8.0, DML's in non-transactional engines like Blackhole engine, do not fail when `legacy_global_read_lock_mode=0` and read-only is turned on. Specifically, when read-only is turned on, in-flight DMLs do not get blocked at commit time and are able to go through. This means that DML statements can be lost since they are seen as committed, despite read-only being on.

To address this, we can make Blackhole engine transactional.

More context here: https://fb.quip.com/1YNyAPbxMevC
The original Task ticket also has a lot of historical context.

Github pull request: https://github.com/facebook/mysql-5.6/pull/1276

Reviewed By: abhinav04sharma

Differential Revision: D43785777