support multiple conditions for the same field
Summary: Today, bypass engine supports 2 conditions with lower/upper bound, like "where a >= 3 and a <= 5", but doesn't support multiple conditions for the same field like "a >= 3 and a >= 5" or "a <= 3 and a <= 5". Let's support these conditions by skipping unnecessary conditions (eg. skipping a >=3 with the presence of a >= 5, skipping a <= 5 with the presence of a <= 3)

Reviewed By: luqun

Differential Revision: D43900214

fbshipit-source-id: ead4ed4f106574af7b604a63833d06fbe77adae1