Convert LOCK_thd_security_ctx to MDL mutex
Summary:
`LOCK_thd_security_ctx` mutex is held across `show processlist` and queries from `information_schema.processlist`. It is also held when `COM_CHANGE_USER` is executed. This is prone to yielding with held lock, and in case of `COM_CHANGE_USER` it performs the switch between databases. The switch results in moving the connection into the system AC info. As a part of the move the worker enters the TP scheduler, and the held lock is detected. Collecting the stack trace for the held lock takes a long time (up to 1 second) and triggers the stuck scheduler detection. Avoiding this detection will be done in another diff. Nevertheless, the detection moves the worker off the scheduler, and when it enters the system AC info it has to get enqueued in the runnable queue instead of just being admitted directly. This resulted in the assert that failed the `main.plugin` test.

This fix is to replace the mutex with MDL lock which is TP friendly and doesn't trigger the held lock detection. Since this mutex is per `THD` object, the corresponding MDL lock would have to have a unique name that could be constructed by using the `THD` object address. This will also result in many MDL locks inserted in the MDL hash table. While I was able to confirm that the growth of MDL lock objects is bounded as the lock objects are reused, still I observed about 1K objects created resulting in ~500KB of extra memory usage.

Memory usage after running this test script: P700001880.
```
mysql> select * from performance_schema.memory_summary_global_by_event_name where event_name like '%lf_node%';
+----------------------+-------------+------------+---------------------------+--------------------------+----------------+--------------------+-----------------+--------------------------+------------------------------+---------------------------+
| EVENT_NAME           | COUNT_ALLOC | COUNT_FREE | SUM_NUMBER_OF_BYTES_ALLOC | SUM_NUMBER_OF_BYTES_FREE | LOW_COUNT_USED | CURRENT_COUNT_USED | HIGH_COUNT_USED | LOW_NUMBER_OF_BYTES_USED | CURRENT_NUMBER_OF_BYTES_USED | HIGH_NUMBER_OF_BYTES_USED |
+----------------------+-------------+------------+---------------------------+--------------------------+----------------+--------------------+-----------------+--------------------------+------------------------------+---------------------------+
| memory/mysys/lf_node |        1806 |          0 |                   1123632 |                        0 |              0 |               1806 |            1806 |                        0 |                      1123632 |                   1123632 |
+----------------------+-------------+------------+---------------------------+--------------------------+----------------+--------------------+-----------------+--------------------------+------------------------------+---------------------------+
1 row in set (0.49 sec)

# 1806 objects above the baseline of ~900.
```

To avoid this, this fix instead creates just 8 locks and assigns them to `THD` objects based on the thread id, similar to how the thread manager assigns the thread remove mutexes. `COM_CHANGE_USER` is not a frequent operation and having 8 partitions should be enough to avoid contention.

Reviewed By: lth

Differential Revision: D45162083

fbshipit-source-id: f839ae723376fe3a5a58640d67bb747faee3bd9e