Add Write_throttle_lag STATUS VAR
Summary:
Add STATUS VAR and test case that exposes the write throttling lag as
perceived by the primary, based on its configured heuristics.

This will later be funneled through into ODS as one metric to help us
understand the write throttling subsystem. In later changes, we will add
more metrics like the number of throttling actions taken, and the number
of rules currently active (and their combined rates) to get a better
understanding of the auto throttling logic in production.

Reviewed By: george-reynya

Differential Revision: D45504491

fbshipit-source-id: 945a99a16ede0c90a2a76158aaf7cba1367e972b