Enable group by plans for more scenarios
Summary:
This fixes https://bugs.mysql.com/bug.php?id=107535

Given an index key(a, b), the following queries will now be able to use group by loose index scans:
```
select b from t where a = 1 group by b
select distinct b from t where a = 1
```

This also removes ordering requirements on the GROUP BY clause, and removes duplicates as well. For example, group by loose index scan is now considered for `GROUP BY a, b`, `GROUP BY b, a`, `GROUP BY a, b, a` whereas previously, the GROUP BY clause had to follow index order.

As mentioned in the bug report, ideally, any columns functionally dependent on the group columns should qualify as an implicit group, but it's not easy to call into the code that determines that so this was left unimplemented. For example, if we have `PK(a, b)` and `key(a, c, b)`, then a query with `GROUP by a, b` should be able to use the key (a, c, b) because c is functionally dependent on (a, b).

Differential Revision: D46043149

fbshipit-source-id: 60735ebeb85da7b685b623237bf4af790e606d18