Do not specify underlying enum type when there is no need to
A plain 'char' without 'signed' or 'unsigned' may default to either one,
depending on the platform. On x86_64 Linux it is signed, while on ARM Linux it
is unsigned, resulting in a build error with enum class Check_database_decision.
The uses of this enum do not depend on the underlying type so just drop it and
let compiler pick one.

Squash with 295ba21bc021c0758ac4f6bb7c35b54ac5093012