MyRocks: implement attachable transaction support
The storage engine requirement for supporting attachable transaction is being
able to handle vanished THD::ha_data field in the middle of transaction. The
server layer swaps its existing data with a nullptr before calling the SE, and
expects the SE to allocate a new transaction object and to invalidate any cached
info depending on the previous one. Once the server is done with the attachable
transaction, it swaps back the original SE data pointer.

MyRocks already happens to work correctly in this case, so the main change is to
advertise the attachable transaction support by adding
HA_ATTACHABLE_TRX_COMPATIBLE to ha_rocksdb::table_flags return value.

Since the attachable transactions are
AC-NL-RO-RC (AutoCommit-NonLocking-ReadOnly-ReadCommitted) at this point, the
rest of code patch introduces invariants asserting these properties for
Rdb_transcation and child class objects and minor changes such as not setting
lock timeouts as no locks can be taken. Specifically, such transaction is
indicated by m_max_row_locks == 0. This is decided in set_params which now
handles both regular and attachable transactions.

Make m_ddl_transaction a private field with getters and setters so that the
invariants could be checked.

Factor out is_autocommit and get_ha_data_or_null helper functions.

At the same time make cleanups in the touched code:
- make field modified_tables private (trivially)
- remove unused write-only field m_mysql_gtid.
- simplify can_acquire_snapshot_without_conflicts, make it `[[nodiscard]]`.
- make ingest_bulk_load_files take const reference argument.
- remove redundant "virtual" keyword from
  Rdb_transaction_impl::is_writebatch_trx.
- remove unused method Rdb_transaction_impl::has_snapshot

For testing, convert the existing InnoDB attachable_trx.test to an include file
that is shared by both engine tests.

At a later point the attachable transactions may become read-write (if i.e.
mysql.gtid_executed is converted to MyRocks) and read-uncommitted, if this
isolation level will be implemented for the DDSE. In this case the invariant
will have to be split.

Reference:
- Server attachable transaction support:
  https://dev.mysql.com/worklog/task/?id=8003
- InnoDB attachable transaction support:
  https://dev.mysql.com/worklog/task/?id=7828
- The initial InnoDB implementation: 5e9eedb9123d0199f6c42c3e5f0a8cd24260c07c